requirements.txt
aniso8601==9.0.1
attrs==21.4.0
black==22.1.0
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
flasgger==0.9.5
Flask==2.0.2
flask-marshmallow==0.14.0
Flask-RESTful==0.3.9
flask-restx==0.5.1
Flask-SendGrid==0.7.1
Flask-SQLAlchemy==2.5.1
Flask-Twilio==0.0.6
greenlet==1.1.2
idna==3.3
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==4.4.0
loguru==0.6.0
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-sqlalchemy==0.27.0
mistune==2.0.2
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.1
psycopg2-binary==2.9.3
PyJWT==2.3.0
pyrsistent==0.18.1
python-dotenv==0.19.2
python-http-client==3.3.5
pytz==2021.3
PyYAML==6.0
requests==2.27.1
sendgrid==6.9.5
six==1.16.0
SQLAlchemy==1.4.31
starkbank-ecdsa==2.0.3
tomli==2.0.1
twilio==7.5.1
typing-extensions==4.0.1
urllib3==1.26.8
Werkzeug==2.0.3
zipp==3.7.0
